+++ src/main/java/com/qrw/controller/QuestionController.java
@@ -19,7 +19,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(value = "id") Integer id, Model model){
+    public String question(@PathVariable(value = "id") Long id, Model model){
 
         QuestionDTO questionDTO = questionService.getById(id);
         questionService.incView(id);
